modified README.md and test_find_roles.py
Changed the order of a few items on my 'things I'd like to do' list as to better reflect the order I want to work on them.  Added a bit of documentation for test_find_roles.py in the README.md file.  Updated test_find_roles.py's acceptable_roles dic to reflect all the new 'role's in README.md now